backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6

pydantic==2.8.2
pydantic-settings==2.5.2

sqlalchemy==2.0.34
psycopg[binary]==3.2.1

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

boto3==1.34.162
botocore==1.34.162

pillow==10.4.0

torch==2.3.1
torchvision==0.18.1

reportlab==4.2.2
